Embed verify.rs in interpreter.rs
